requirements.txt
altair==5.2.0
attrs==23.1.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
extra-streamlit-components==0.1.60
fastjsonschema==2.19.0
fonttools==4.46.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
nbformat==5.9.2
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.1
pyarrow==14.0.2
pydeck==0.8.1b0
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
smmap==5.0.1
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
